% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vrwatch}
\alias{vrwatch}
\alias{vrwatch-package}
\title{Log the loading and unloading of R packages}
\description{
Log the loading and unloading of R packages
}
\section{Introduction}{

\strong{vrwatch} is a tool that logs the loadings and unloadings of R packages
on a system. Its goal is to measure which packages are used, and how
often.
}

\section{Requirements}{

\itemize{
\item A version of R. Any version later than or equal to 3.0.0 is sufficient.
\item To log to the system log, a logger daemon, and the \code{logger} command
line tool must be available. The \strong{vrwatch} RPM package
configures \code{rsyslog} (http://www.rsyslog.com/), but in general any
system compatible with \code{logger} can be used. \code{rsyslog} is standard and
installed by default on most Linux distributions, including Debian
Linux, Ubuntu Linux, CentOS and RedHat Linux.

The \strong{vrwatch} RPM package sets up \code{rsyslog} to log \strong{vrwatch}
events to \code{/var/log/vrwatch.log}. If an alternative logging system is
used then these events go to the regular generic system log file by
default.
\item The \strong{vrwatch} R package needs to be loaded in every R session that
is to be logged. The \strong{vrwatch} RPM package sets up the
system-wide R profile in \code{$R_HOME/etc/Rprofile.site} to load
\strong{vrwatch} at the beginning of every R session.
}
}

\section{Installation}{

\strong{vrwatch} is an R package, and you can install it as a regular R package.
We suggest to install it via the supplied RPM package, whenever
possible; to set up the \code{rsyslog} system logger, and the automatic
loading of \strong{vrwatch} into every R session via a system-wide R profile.

For platforms unsupported by the supplied RPM package,
\strong{vrwatch} can be installed as a regular R package. The filtering of
the \strong{vrwatch} events to a separate log file can be set up by the
system administrator. \strong{vrwatch} events are logged with the \strong{vrwatch}
tag, so the following \code{rsyslog} configuration file can be used:\preformatted{:syslogtag, isequal, "VRWATCH:" /var/log/vrwatch.log
& ~
}

Save this to \code{/etc/rsyslog/23-vrwatch.conf} and restart the \code{rsyslog}
daemon.

To check that \strong{vrwatch} is installed and working properly, start R,
and check that the \strong{vrwatch} package is loaded:\preformatted{loadedNamespaces()
#> [1] "graphics"  "utils"     "grDevices" "stats"
#> [5] "datasets"  "vrwatch"   "methods"   "base"
}

Then load an R package (e.g. \strong{splines}, which is not loaded by default,
but included on almost all R installations: \code{library(splines)}).
Try unloading a package as well, e.g. \code{unloadNamespace("splines")}.
Then verify that the package loadings are logged to the log file,
which is \code{/var/log/vrwatch.log}, unless configured otherwise.
}

\section{Log entries}{

\strong{vrwatch} log entries look similar to the following example. Note that
there might be some differences depending on the logging configuration
of the system.\preformatted{Nov  3 10:27:01 centos6 VRWATCH: 21735 LOAD   splines /usr/lib64/R/library/splines
Nov  3 10:27:04 centos6 VRWATCH: 21735 UNLOAD splines /usr/lib64/R/library/splines
}
\itemize{
\item The first three columns are a timestamp, in the regular log format
\item The fourth column is the hostname
\item The fifth column is a log tag, which is always \code{VRWATCH:}
\item The six column is the process id of the R process
\item The seventh column is \code{LOAD} or \code{UNLOAD}, the type of the event
\item The eighth column is the name of the R package that was loaded or
unloaded
\item The last column is the directory where the R package is installed
on the system
}
}

\section{Logging to a system log}{

\strong{vrwatch} logs to the system log via the \code{logger} command line program
by default. To avoid this, set the \code{VRWATCH_NOSYSLOG} environment
variable to any non-empty value, before loading the \strong{vrwatch} R package.
}

\section{Logging to a file}{

In addition or instead the system log, \strong{vrwatch} can log to an
arbitrary log file. For this the \code{VRWATCH_LOGFILE} environment variable
must be set to the name of the file to be logged to, before loading the
\strong{vrwatch} package.
}

\section{Impact on the system}{

\strong{vrwatch} traces (see \code{\link[base:trace]{base::trace()}}) the \code{\link[base:loadNamespace]{base::loadNamespace()}} and
the \code{\link[base:unloadNamespace]{base::unloadNamespace()}} base R functions, and inserts an extra
function call into them, to do the logging.

The impact of \strong{vrwatch} on the system should be very minimal:
\itemize{
\item Loading and unloading package will be slightly slower when \strong{vrwatch}
is loaded. In our measurements the difference in running time was
less than 0.01 seconds, so this is only an issue for code that
repeatedly loads and unloads R packages many times, and this is
very atypical.
\item The \strong{vrwatch} package must be loaded at all times to perform the
logging, so code that queries and manipulates the set of loaded
packages can be potentially impacted. Again, this is very atypical.
}

In case of unwanted system impact, unload the \strong{vrwatch} package:\preformatted{unloadNamespace("vrwatch")
}
}

\author{
Mango Solutions http://mango-solutions.com
}
